backend/requirements.txt
# Web Framework & API
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database & Storage
prisma>=0.14.0
prisma-client-python>=0.14.0
sqlalchemy>=2.0.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Google Workspace Integration
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# NLP, ML & OCR (AI Copilot Engine)
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
spacy>=3.7.0
nltk>=3.8.1
sentence-transformers>=2.4.0
torch>=2.1.0
transformers>=4.36.0
joblib>=1.3.2
tqdm>=4.66.0

# Document Parsing & Processing
pdfplumber>=0.10.0
PyMuPDF>=1.23.8
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0

# Utilities & Networking
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.27.0

# Testing & Quality Assurance
pytest>=8.0.0
pytest-asyncio>=0.23.0
